refactor(cli): move the pmtiles group into its own module
Cuts `gpio pmtiles` out of cli/main.py into
geoparquet_io/cli/commands/pmtiles.py, unchanged. The group was the last
block in main.py, above `if __name__ == "__main__"`, and is a clean cut: no
shared helpers, no test patch sites to repoint.

tests/data/cli_surface.json is byte-identical.
